requirements.txt
PySide6>=6.10.1
xlsxwriter>=3.2.9
PyYAML>=6.0.3
reportlab>=4.0.0
